# server/abl/taxPayer/deleteAbl.js
const Ajv = require("ajv");
const addFormats = require("ajv-formats").default;
const taxPayerDao = require("../../dao/taxPayer-dao.js");
const invoiceDao = require("../../dao/invoice-dao.js");
const openApiSchema = require("../../schema/openapi/schema.json");

const ajv = new Ajv();
addFormats(ajv);

const pathItem = openApiSchema.paths["/taxpayers/taxpayer/{id}"];
const operation = pathItem.delete;

const allParameters = [
  ...(pathItem.parameters || []),
  ...(operation.parameters || []),
];

const properties = {};
const required = [];

allParameters.forEach((param) => {
  properties[param.name] = param.schema;
  if (param.required && !required.includes(param.name)) {
    required.push(param.name);
  }
});

const schema = {
  type: "object",
  properties,
  required,
  additionalProperties: false,
};

const validate = ajv.compile(schema);

async function DeleteAbl(req, res) {
  try {
    const reqParams = req.params;

    // validate input
    const valid = validate(reqParams);
    if (!valid) {
      return res.status(400).json({
        code: "requestIsNotValid",
        message: "Request is not valid",
        validationError: ajv.errors,
      });
    }

    const existingTaxPayer = await taxPayerDao.get(reqParams.id);
    if (!existingTaxPayer) {
      return res.status(404).json({
        code: "taxPayerNotFound",
        message: `Tax payer with id ${reqParams.id} not found`,
      });
    }

    // Referential integrity check: are there invoices assigned to this tax payer?
    const { itemList: invoiceList } = await invoiceDao.list({ taxPayerId: reqParams.id });
    if (invoiceList && invoiceList.length > 0) {
      return res.status(400).json({
        code: "taxPayerHasInvoices",
        message: "Tax payer has related invoices and cannot be deleted",
      });
    }

    await taxPayerDao.remove(reqParams.id);

    res.status(200).json({});
  } catch (error) {
    console.error(error);
    res.status(500).json({ message: error.message });
  }
}

module.exports = DeleteAbl;