# server/schema/openapi/schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EasyVAT REST API",
    "description": "API for managing taxpayers and their issued or received invoices, including tax exports.",
    "version": "1.3.0"
  },
...
  "paths": {
...
    "/taxpayers/taxpayer/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "TaxPayer"
        ],
        "summary": "Get tax payer details by ID",
        "operationId": "getTaxPayerById",
        "responses": {
          "200": {
            "description": "Successful retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxPayer"
                }
              }
            }
          },
          "400": {
            "description": "Request is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorValidation"
                }
              }
            }
          },
          "404": {
            "description": "Tax payer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TaxPayer"
        ],
        "summary": "Update an existing taxpayer",
        "operationId": "updateTaxPayer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxPayer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxPayer"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorValidation"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TaxPayer"
        ],
        "summary": "Delete a tax payer",
        "operationId": "deleteTaxPayer",
        "responses": {
          "204": {
            "description": "Deleted successfully"
          }
        }
      }
    },
...
    "/invoices/invoice/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "Invoice"
        ],
        "summary": "Get invoice details",
        "operationId": "getInvoiceById",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Invoice"
        ],
        "summary": "Update invoice",
        "operationId": "updateInvoice",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Invoice"
        ],
        "summary": "Delete invoice",
        "operationId": "deleteInvoice",
        "responses": {
          "204": {
            "description": "Deleted successfully"
          }
        }
      }
    },
...
    "/export/dp": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export VAT return (DPH) to XML",
        "operationId": "exportDphXml",
        "parameters": [
          {
            "name": "taxPayerId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "XML generated",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/export/kh": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export VAT control statement (KH) to XML",
        "operationId": "exportKhXml",
        "parameters": [
          {
            "name": "taxPayerId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "XML generated",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TaxPayer": {
        "type": "object",
        "required": [
          "type",
          "vatId",
          "street",
          "descriptiveNumber",
          "city",
          "postalCode",
          "email",
          "phoneNumber"
        ],
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "type": {
            "type": "string",
            "enum": [
              "individual",
              "company"
            ]
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "vatId": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "descriptiveNumber": {
            "type": "string"
          },
          "referenceNumber": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      },
      "Invoice": {
        "type": "object",
        "required": [
          "taxPayerId",
          "type",
          "invoiceNumber",
          "vatId",
          "name",
          "taxableDate",
          "price",
          "vatValue"
        ],
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "taxPayerId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "issued",
              "received"
            ]
          },
          "invoiceNumber": {
            "type": "string"
          },
          "vatId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "taxableDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "vatValue": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "readOnly": true
          },
          "message": {
            "type": "string",
            "readOnly": true
          }
        }
      },
      "ErrorValidation": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "readOnly": true
          },
          "message": {
            "type": "string",
            "readOnly": true
          },
          "validationError": {
            "type": "string",
            "readOnly": true
          }
        }
      }
    }
  }
}